Move `Widgets.Mediator` to its own file
